requirements.txt
pip
kafka-python
argparse
sentence_transformers
tensorflow
tensorflow_hub
torch
sklearn
time